增加开户功能

+++ model/src/main/java/com/zwl/model/po/UserAccount.java
@@ -10,7 +10,7 @@ public class UserAccount {
 
     private String userId;
 
-    private Double balance;
+    private Integer balance;
 
     private Date createTime;
